requirements.txt
flask
gunicorn
python_binance